+++ programs/store/src/instructions/exchange/mod.rs
@@ -28,6 +28,9 @@ pub mod shift;
 /// Execute shift.
 pub mod execute_shift;
 
+/// Update closed state.
+pub mod update_closed;
+
 pub use deposit::*;
 pub use execute_deposit::*;
 pub use execute_order::*;
@@ -37,6 +40,7 @@ pub use order::*;
 pub use position_cut::*;
 pub use shift::*;
 pub use update_adl::*;
+pub use update_closed::*;
 pub use withdrawal::*;
 
 use crate::CoreError;