+++ programs/store/src/states/market/mod.rs
@@ -93,7 +93,8 @@ pub struct Market {
     /// Bump Seed.
     pub(crate) bump: u8,
     flags: MarketFlagContainer,
-    padding: [u8; 13],
+    padding: [u8; 5],
+    closed_state_updated_at: i64,
     #[cfg_attr(feature = "serde", serde(with = "serde_bytes"))]
     name: [u8; MAX_NAME_LEN],
     pub(crate) meta: MarketMeta,
@@ -287,8 +288,10 @@ impl Market {
     /// Set whether the market is closed.
     ///
     /// Return previous value.
-    pub fn set_closed(&mut self, closed: bool) -> bool {
-        self.set_flag(MarketFlag::Closed, closed)
+    pub(crate) fn set_closed(&mut self, closed: bool) -> Result<bool> {
+        let clock = Clock::get()?;
+        self.closed_state_updated_at = clock.unix_timestamp;
+        Ok(self.set_flag(MarketFlag::Closed, closed))
     }
 
     /// Get pool of the given kind.